pyre-jit-trace: extract pcmap/audit helpers to diag.rs
Move the jitcode-pc -> python-pc inversion helpers
(python_pc_for_jitcode_pc + floor-boundary), skip_python_trivia_forward,
resolve_parent_resume_py_pc, the in-flight FOR_ITER body resolvers, and
the surviving RECIPE_RESULTCOLOR audit probe into diag.rs.
